requirements.txt
pandas>=2.0.0
openpyxl>=3.1.0
pyyaml>=6.0
svgwrite>=1.4.0
reportlab>=4.0.0
svglib>=1.5.0
click>=8.0.0
rich>=13.0.0
pydantic>=2.0.0
pytest>=7.0.0
anthropic>=0.30.0
flask>=2.0.0
pdf2image>=1.16.0
pytesseract>=0.3.10
Pillow>=10.0.0
PyPDF2>=3.0.0
